+++ CHANGES.md
@@ -1,4 +1,9 @@
 
+## 7.0.0 (TBD)
+
+* add `_tile_matrices_idx: Dict[str, int]` private attribute to improve `matrices` lookup
+* change `_ul`, `ul`, `_lr` and `lr` method to only accept `morecantile.Tile` object **breaking change**
+
 ## 6.0.0 (2024-10-16)
 
 * remove `_geographic_crs` private attribute in `TileMatrixSet` model **breaking change**